requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
jinja2==3.1.6
weaviate-client==4.20.4
pypdf==6.9.2
python-multipart==0.0.24
langchain-anthropic==1.4.0
langchain-text-splitters==1.1.1
yagmail==0.15.293
Pillow==11.3.0
google-genai
anam
livekit-agents[silero,google]>=1.4
livekit-plugins-anam
livekit-plugins-noise-cancellation
